+++ src/test/java/io/nanofaas/containerd/internal/TasksServiceImplTest.java
@@ -4,7 +4,6 @@ import io.grpc.Status;
 import io.grpc.inprocess.InProcessChannelBuilder;
 import io.grpc.inprocess.InProcessServerBuilder;
 import io.grpc.stub.StreamObserver;
-import io.nanofaas.containerd.Signal;
 import org.junit.jupiter.api.Test;
 
 import java.util.concurrent.atomic.AtomicInteger;